# src/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This ios Intellidrone API designed by [Singleton](http://singleton.com.ar).You\
    \  can find out more about Swagger at [Singleton Web Page](http://singleton.com.ar)\
    \ or on [Intellidrone Web Page](http://robotagro.com)."
  version: "1.0.0"
  title: "Intellidrone API"
  termsOfService: "http://web.robotagro.com/api//terms"
  contact:
    name: "Equipo de Soporte"
    url: "http://singleton.com.ar/soporte"
    email: "soporte@singleton.com.ar"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/"
...
schemes:
- "http"
- "https"
paths:
  /Lotes:
    get:
      tags:
      - "Lotes"
      summary: "Get all lotes."
      description: "Get all lotes."
      operationId: "getLotes"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "skip"
        in: "query"
        description: "number of item to skip (not included the first)"
        required: true
        type: "integer"
        format: "int32"
      - name: "limit"
        in: "query"
        description: "max records to return (included the last)"
        required: true
        type: "integer"
        format: "int32"
      - name: "orderBy"
        in: "query"
        description: "order by property. Put '-' at the beginning to descendant order."
        required: false
        type: "string"
      - name: "filter"
        in: "query"
        description: "filter data. NOT SUPPORTED YET."
        required: false
        type: "string"
      - name: "userId"
        in: "query"
        description: "id of user. Only for admin users."
        required: false
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Request OK."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Lotes"
        204:
          description: "Response is empty."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        400:
          description: "Some parameters are missing or badly entered."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "string"
              format: "date-time"
              description: "The UTC date/time at which the current rate limit window\
                \  resets."
          schema:
            type: "string"
        404:
          description: "Entity not found."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Illegal input for operation."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
      security:
      - appToken: []
      - userToken: []
      x-swagger-router-controller: "Lotes"
    post:
      tags:
      - "Lotes"
      summary: "Add one lote."
      description: "Add one lote."
      operationId: "addLote"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "Lote"
        required: true
        schema:
          $ref: "#/definitions/Lotes"
      responses:
        200:
          description: "Request OK."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Lotes"
        204:
          description: "Response is empty."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        400:
          description: "Some parameters are missing or badly entered."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "string"
              format: "date-time"
              description: "The UTC date/time at which the current rate limit window\
                \ / resets."
          schema:
            type: "string"
        404:
          description: "Entity not found."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Illegal input for operation."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
      security:
      - appToken: []
      - userToken: []
      x-swagger-router-controller: "Lotes"
    put:
      tags:
      - "Lotes"
      summary: "Edit one lote."
      description: "Edit one lote."
      operationId: "editLote"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "Lote"
        required: true
        schema:
          $ref: "#/definitions/Lotes"
      responses:
        200:
          description: "Request OK."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Lotes"
        204:
          description: "Response is empty."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        400:
          description: "Some parameters are missing or badly entered."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "string"
              format: "date-time"
              description: "The UTC date/time at which the current rate limit window\
                \  resets."
          schema:
            type: "string"
        404:
          description: "Entity not found."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Illegal input for operation."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
      security:
      - appToken: []
      - userToken: []
      x-swagger-router-controller: "Carriers"
  /Lotes/{id}:
    get:
      tags:
      - "Lotes"
      summary: "Get one lote."
      description: "Get one lote."
      operationId: "getLoteById"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id"
        in: "path"
        description: "id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Request OK."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Lotes"
        204:
          description: "Response is empty."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        400:
          description: "Some parameters are missing or badly entered."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "string"
              format: "date-time"
              description: "The UTC date/time at which the current rate limit window\
                \  resets."
          schema:
            type: "string"
        404:
          description: "Entity not found."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Illegal input for operation."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
      security:
      - appToken: []
      - userToken: []
      x-swagger-router-controller: "Lotes"
    delete:
      tags:
      - "Lotes"
      summary: "Delete one lote."
      description: "Delete one lote."
      operationId: "deleteLote"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id"
        in: "path"
        description: "id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Request OK."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/DeletedResponse"
        204:
          description: "Response is empty."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        400:
          description: "Some parameters are missing or badly entered."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "string"
              format: "date-time"
              description: "The UTC date/time at which the current rate limit window\
                \  resets."
          schema:
            type: "string"
        404:
          description: "Entity not found."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Illegal input for operation."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
      security:
      - appToken: []
      - userToken: []
      x-swagger-router-controller: "Lotes"
...
securityDefinitions:
  appToken:
    type: "apiKey"
    name: "x_app_id"
    in: "header"
  userToken:
    type: "apiKey"
    name: "x_user_key"
    in: "header"
definitions:
  Lotes:
    type: "object"
    properties:
      _id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      latitude:
        type: "number"
        format: "float"
      longitude:
        type: "number"
        format: "float"
    example:
      latitude: 6.0274563
      name: "Campo Baires"
      id: 0
      longitude: 1.4658129
...
  Error:
    type: "object"
    required:
    - "message"
    properties:
      code:
        type: "string"
      message:
        type: "string"
      data:
        type: "string"
    example:
      code: "code"
      data: "data"
      message: "message"
  DeletedResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
    example:
      id: 0